backend/requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
google-genai>=1.0.0
PyJWT>=2.8.0
